+++ phoenix/monitor/views/__init__.py
@@ -8,7 +8,8 @@ logger = logging.getLogger(__name__)
 
 @subscriber(JobStarted)
 def notify_job_started(event):
-    event.request.session.flash("Job added to task queue. Please wait ...", queue='info')
+    event.request.session.flash(
+        '<h4><img src="/static/phoenix/img/ajax-loader.gif"></img> Job Created. Please wait ...</h4>', queue='warning')
 
 
 @subscriber(JobFinished)
@@ -20,5 +21,3 @@ def notify_job_finished(event):
         logger.warn("job %s failed.", event.job.get('title'))
         # logger.warn("status = %s", event.job.get('status'))
         # event.request.session.flash("Job <b>{0}</b> failed.".format(event.job.get('title')), queue='danger')
-
-