+++ phoenix/monitor/views/details.py
@@ -12,14 +12,10 @@ class Details(MyView):
         super(Details, self).__init__(
             request, name='monitor_details', title='Details')
 
-    def breadcrumbs(self):
-        breadcrumbs = super(Details, self).breadcrumbs()
-        breadcrumbs.append(dict(route_path=self.request.route_path('monitor'), title='Monitor'))
-        breadcrumbs.append(dict(route_path='', title=self.title))
-        return breadcrumbs
-
     @view_config(route_name='monitor_details', renderer='../templates/monitor/details.pt')
     def view(self):
         tab = self.request.matchdict.get('tab')
         job_id = self.request.matchdict.get('job_id')
-        return dict(active=tab, job_id=job_id)
+        collection = self.request.db.jobs
+        job = collection.find_one({'identifier': job_id})
+        return dict(active=tab, job_id=job_id, status=job['status'])