+++ phoenix/tasks/execute.py
@@ -7,7 +7,7 @@ from owslib.wps import WebProcessingService
 
 from phoenix.db import mongodb
 from phoenix.events import JobFinished
-from phoenix.tasks.utils import wps_headers, log, log_error, add_job, wait_secs
+from phoenix.tasks.utils import wps_headers, save_log, add_job, wait_secs
 
 from celery.utils.log import get_task_logger
 logger = get_task_logger(__name__)
@@ -24,6 +24,7 @@ def execute_process(self, url, service_name, identifier, inputs, outputs, async=
         service_name=service_name,
         process_id=identifier,
         is_workflow=False,
+        async=async,
         caption=caption)
 
     try:
@@ -40,11 +41,16 @@ def execute_process(self, url, service_name, identifier, inputs, outputs, async=
 
         num_retries = 0
         run_step = 0
-        while execution.isNotComplete() or not run_step:
+        while execution.isNotComplete() or run_step == 0:
             if num_retries >= 5:
                 raise Exception("Could not read status document after 5 retries. Giving up.")
             try:
                 execution.checkStatus(sleepSecs=wait_secs(run_step))
+                # TODO: fix owslib response object
+                if isinstance(execution.response, etree._Element):
+                    job['response'] = etree.tostring(execution.response)
+                else:
+                    job['response'] = execution.response.encode('UTF-8')
                 job['status'] = execution.getStatus()
                 job['status_message'] = execution.statusMessage
                 job['progress'] = execution.percentCompleted
@@ -54,14 +60,13 @@ def execute_process(self, url, service_name, identifier, inputs, outputs, async=
                 if execution.isComplete():
                     job['finished'] = datetime.now()
                     if execution.isSucceded():
+                        logger.debug("job succeded")
                         job['progress'] = 100
-                        log(job)
                     else:
+                        logger.debug("job failed.")
                         job['status_message'] = '\n'.join(error.text for error in execution.errors)
                         for error in execution.errors:
-                            log_error(job, error)
-                else:
-                    log(job)
+                            save_log(job, error)
             except:
                 num_retries += 1
                 logger.exception("Could not read status xml document for job %s. Trying again ...", self.request.id)
@@ -69,13 +74,15 @@ def execute_process(self, url, service_name, identifier, inputs, outputs, async=
                 logger.debug("update job %s ...", self.request.id)
                 num_retries = 0
                 run_step += 1
+            finally:
+                save_log(job)
                 db.jobs.update({'identifier': job['identifier']}, job)
     except Exception as exc:
         logger.exception("Failed to run Job")
         job['status'] = "ProcessFailed"
-        job['status_message'] = "Failed to run Job. %s" % exc.message
+        job['status_message'] = "Error: {0}".format(exc.message)
     finally:
-        log(job)
+        save_log(job)
         db.jobs.update({'identifier': job['identifier']}, job)
 
     registry.notify(JobFinished(job))